requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
pydantic
google-genai
requests